requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
mysqlclient
alembic
python-dotenv
passlib[bcrypt]
python-jose
pytest
pytest-asyncio
httpx
pymysql
pydantic-settings
